+++ BambooTracker/io/instrument_io.hpp
@@ -59,7 +59,7 @@ private:
 			FMEnvelopeParameter param, size_t instMemCsr, std::shared_ptr<InstrumentsManager>& instManLocked,
 			BinaryContainer& ctr, InstrumentFM* inst, int idx, uint32_t version);
 	static size_t getPropertyPositionForBTB(const BinaryContainer& propCtr, uint8_t subsecType, uint8_t index);
-	static int convertTFIVGMDT(int dt);
+	static int convertDTInTFIVGIDMP(int dt);
 
 
 private: